+++ R/CFDataset.R
@@ -77,6 +77,7 @@ CFDataset <- R6::R6Class("CFDataset",
         cat("\nAxes:\n")
         axes <- do.call(rbind, lapply(self$root$axes(), function(a) a$brief()))
         axes <- lapply(axes, function(c) if (all(c == "")) NULL else c)
+        if (all(axes$group == "/")) axes$group <- NULL
         axes <- unique(as.data.frame(axes[lengths(axes) > 0L]))
         print(.slim.data.frame(axes, 50L), right = FALSE, row.names = FALSE)